+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [unreleased]
 
+### Changed
+
+- Dropped Python 3.8 support.
+
+### Removed
+
+- `black` and `isort` in favor of `ruff`.
+
 ## [0.6.4] - 2024-10-01
 
 ### Added